i#7303: Integrate Drsyscall into DynamoRIO. (#7396)
The following changes are made to integrate DrSyscall into DynamoRIO:

- Update copyright.
- Remove drsyscall.dox.
- Add drcontext as the first parameter to LOG() to be consistent with
DynamoRIO.
- Add drcontext as the first parameter to name2num_entry_add(),
name2num_record(), wingdi_add_usercall(), and
wingdi_get_secondary_syscall_num() to pass it along to LOG().
-  Change the include file path for wininc/* to drmf/wininc/*.
- Remove rlimit64 declaration from linux_defines.h.
- Update drsyscall/CMakeLists.txt to include files in framework and
common.
- Copy the DRMF_VERSION and set_library_version logic from
drmemory/CMakeLists.txt to ext/drsyscall/CMakeList.txt.
- Skip Alpine, ANDROID and RISCV64 build in drsycall/CMakeLists.txt
- Change make/resources.rc to core/win32/resources.rc in
drsyscall/CMakeLists.txt.

Issue: #7303

+++ ext/drsyscall/pdb2sysfile.cpp
@@ -1,5 +1,5 @@
 /* **********************************************************
- * Copyright (c) 2011-2024 Google, Inc.  All rights reserved.
+ * Copyright (c) 2011-2025 Google, Inc.  All rights reserved.
  * Copyright (c) 2003-2008 VMware, Inc.  All rights reserved.
  * **********************************************************/
 
@@ -46,8 +46,8 @@
 #include "dr_frontend.h"
 #include "droption.h"
 #include "drsyms.h"
-#include "../common/utils.h" /* only for BUFFER*, DIRSEP */
-#include "../drsyscall/drsyscall.h"
+#include "utils.h" /* only for BUFFER*, DIRSEP */
+#include "drsyscall.h"
 
 #include <assert.h>
 #include <algorithm>
@@ -426,7 +426,7 @@ drsys_find_sysnum_libs(OUT char **sysnum_lib_paths,
     /* Next, get the count of dlls that exist on this machine
      * (win32u.dll and kernelbase.dll do not exist on older Windows).
      */
-    int count = 0;
+    size_t count = 0;
     int dll_readable[NUM_SYSCALL_DLLS];
     char buf[MAXIMUM_PATH];
     for (int i = 0; i < NUM_SYSCALL_DLLS; ++i) {
@@ -479,7 +479,7 @@ fetch_symbols(const char **sysnum_lib_paths, size_t num_sysnum_libs,
         drfront_set_symbol_search_path(symsrv_dir) != DRFRONT_SUCCESS)
         NOTIFY(0, "WARNING: Can't set symbol search path. Symbol lookup may fail." NL);
 
-    for (int i = 0; i < num_sysnum_libs; ++i) {
+    for (size_t i = 0; i < num_sysnum_libs; ++i) {
         char pdb_path[MAXIMUM_PATH];
         drfront_status_t sc = DRFRONT_SUCCESS;
         /* Sometimes there are transient errors on the symbol server side so we
